shishufied: Add our Shishu Immensity
A bit of Uppercase and some rounded icons, nothing else than that. And a darker settings background and black compared to Nights.
Rn it's missing some extra stuff i really wanted to do but adding uppercase to material.smalltext would be so lazily done and people may hate it.
But anyways, thanks to Altan KRK for the inspiration on PB theme, Quetzal dev for inspiration and help with some exposures and substratum team for making that engine and making everything so easy to test